Intermediate fixation
